Issue#190: EntitySpecificationService with all query configs moved to home-application module
