+++ home-application/src/main/java/com/softserveinc/ita/homeproject/application/api/CommonApi.java
@@ -9,7 +9,7 @@ import javax.ws.rs.core.UriInfo;
 import com.softserveinc.ita.homeproject.application.mapper.HomeMapper;
 import com.softserveinc.ita.homeproject.application.service.QueryApiService;
 import com.softserveinc.ita.homeproject.homeservice.dto.BaseDto;
-import com.softserveinc.ita.homeproject.homeservice.query.EntitySpecificationService;
+import com.softserveinc.ita.homeproject.application.service.impl.EntitySpecificationService;
 import com.softserveinc.ita.homeproject.model.BaseReadView;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
