+++ home-application/src/main/java/com/softserveinc/ita/homeproject/application/config/rsql/RSQLConfig.java
@@ -6,7 +6,7 @@ import java.util.stream.Collectors;
 import javax.annotation.PostConstruct;
 
 import com.softserveinc.ita.homeproject.homedata.entity.BaseEntity;
-import com.softserveinc.ita.homeproject.homeservice.query.QueryParamEnum;
+import com.softserveinc.ita.homeproject.application.service.QueryParamEnum;
 import io.github.perplexhub.rsql.RSQLCommonSupport;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
