+++ home-application/src/main/java/com/softserveinc/ita/homeproject/application/config/rsql/RSQLEndpointConfig.java
@@ -3,8 +3,8 @@ package com.softserveinc.ita.homeproject.application.config.rsql;
 import java.util.Map;
 
 import com.softserveinc.ita.homeproject.homedata.entity.BaseEntity;
-import com.softserveinc.ita.homeproject.homeservice.query.QueryConfig;
-import com.softserveinc.ita.homeproject.homeservice.query.QueryParamEnum;
+import com.softserveinc.ita.homeproject.application.service.QueryConfig;
+import com.softserveinc.ita.homeproject.application.service.QueryParamEnum;
 
 public interface RSQLEndpointConfig<T extends BaseEntity, Q extends QueryParamEnum> {
