+++ home-application/src/main/java/com/softserveinc/ita/homeproject/application/config/rsql/UserRSQLEndpointConfig.java
@@ -4,9 +4,9 @@ import java.util.EnumMap;
 import java.util.Map;
 
 import com.softserveinc.ita.homeproject.homedata.entity.User;
-import com.softserveinc.ita.homeproject.homeservice.query.QueryConfig;
-import com.softserveinc.ita.homeproject.homeservice.query.impl.UserQueryConfig;
-import com.softserveinc.ita.homeproject.homeservice.query.impl.UserQueryConfig.UserQueryParamEnum;
+import com.softserveinc.ita.homeproject.application.service.QueryConfig;
+import com.softserveinc.ita.homeproject.application.service.impl.UserQueryConfig;
+import com.softserveinc.ita.homeproject.application.service.impl.UserQueryConfig.UserQueryParamEnum;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
