+++ home-application/src/main/java/com/softserveinc/ita/homeproject/application/service/QueryParamEnum.java
@@ -1,4 +1,4 @@
-package com.softserveinc.ita.homeproject.homeservice.query;
+package com.softserveinc.ita.homeproject.application.service;
 
 public interface QueryParamEnum {
